[VL-1] FilmReviewDao and test is complete

+++ video-library-adapter/jdbc-adapter/src/main/java/it/proconsole/library/video/adapter/jdbc/model/Genre.java
@@ -2,5 +2,5 @@ package it.proconsole.library.video.adapter.jdbc.model;
 
 import it.proconsole.library.video.core.model.GenreEnum;
 
-public record Genre(int id, GenreEnum value) {
+public record Genre(Long id, GenreEnum value) {
 }
